cbt-backend/requirements.txt
fastapi
uvicorn
google-genai
openpyxl
python-dotenv
groq